feat(management): 报告

+++ myopia-business/business-api/management-api/src/main/java/com/wupol/myopia/business/api/management/domain/dto/report/vision/common/Outline.java
@@ -47,12 +47,12 @@ public class Outline {
     /**
      * 合计人数
      */
-    private Integer studentTotal;
+    private Long studentTotal;
 
     /**
      * 本次未筛
      */
-    private Integer unScreening;
+    private Long unScreening;
 
     /**
      * 无效人数
